Switch to using browsers built-in support for gzip decompression
Gets rid of the dependency on the external pako library.

The pasted save string from the in game code "copysave" is never compressed so the dependency on zlib is also removed including the local copy of those packages.

Unified the "clean up" of the save string coming from both the pasted version and the uploaded file instead of duplicating it.

package-lock.json had a bunch of extra dependencies and other weird mismatches with package.json that are also fixed.