Replace old and unmaintained notification dependency with native browser API
"react-push-notification" is an old library with a npm peer dependency on React 16. That causes issues since the project depends on React 18 requiring the flag --legacy-peer-deps to override this requirement when running npm install.

The library, in addition to the above problem, has a much broader scope than what is required by the project. It is only used to show a native notification for which there is a widely supported browser API.

Finally, there was a third problem with the way permission to show notifications was requested. These days it can only be requested in response to a direct user interaction (like a click handler). Since this request was made by the library as it is trying to show the notification, and the library is called by the timer finishing. This is not in response to a user action so the browser will not prompt the end user for notification permission.