Optimize the calculations for required reincarnation XP for a given level
By pre-calculating and storing the fixed bonuses from cow shop and expedition shop the calculations for each individual level is drastically sped up.
Since it tries to calculate the total required reincarnation XP to ascend it does this calculation for possibly hundred of thousands of levels - on each render.

Basic profiling in a browser shows about 40-50% less time spent on page load.